requirements.txt
flask>=2.0.0
openai>=0.27.0
python-dotenv>=0.19.0
apscheduler>=3.8.0
requests>=2.26.0